+++ Softeq.XToolkit.WhiteLabel.Tests/ViewModels/PaginationSearchViewModelBaseTests/PaginationSearchViewModelBaseTests.cs
@@ -2,14 +2,12 @@
 // http://www.softeq.com
 
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using NSubstitute;
 using Softeq.XToolkit.Common.Threading;
 using Softeq.XToolkit.WhiteLabel.Model;
 using Softeq.XToolkit.WhiteLabel.Tests.Stubs;
-using Softeq.XToolkit.WhiteLabel.ViewModels;
 using Xunit;
 
 namespace Softeq.XToolkit.WhiteLabel.Tests.ViewModels.PaginationSearchViewModelBaseTests
@@ -107,30 +105,6 @@ namespace Softeq.XToolkit.WhiteLabel.Tests.ViewModels.PaginationSearchViewModelB
             await _dataLoader
                 .Received()
                 .Invoke(Arg.Is(LastSearchQuery), Arg.Any<int>(), Arg.Any<int>());
-        }
-    }
-
-    internal class TestPaginationSearchViewModelBaseTests : PaginationSearchViewModelBase<string, string>
-    {
-        private readonly Func<string, int, int, Task<PagingModel<string>>> _dataLoader;
-
-        public TestPaginationSearchViewModelBaseTests(Func<string, int, int, Task<PagingModel<string>>> dataLoader)
-        {
-            _dataLoader = dataLoader;
-        }
-
-        public int TestSearchDelay => SearchDelay;
-
-        protected override int SearchDelay => 50;
-
-        protected override IList<string> MapItemsToViewModels(IList<string> models)
-        {
-            return models;
-        }
-
-        protected override Task<PagingModel<string>> LoadAsync(string query, int pageNumber, int pageSize)
-        {
-            return _dataLoader(query, pageNumber, pageSize);
         }
     }
 }
